+++ storage/node/node-api/accounts.js
@@ -8,9 +8,9 @@ const accounts = require("../accounts");
 const Account = require('../account');
 const authorize = require("../authorize");
 const Roles = require("../roles");
-const logger = require("../../utils/logger");
+const { logger } = require('@dictadata/lib')
 const { StorageError } = require("@dictadata/storage-junctions/types");
-const { objCopy } = require("@dictadata/storage-junctions/utils");
+const { objCopy } = require("@dictadata/lib");
 
 /**
  * accounts routes