+++ storage/node/node-api/status.js
@@ -7,7 +7,7 @@ const express = require("express");
 const authorize = require('../authorize');
 const Roles = require('../roles');
 const config = require('../config');
-const logger = require('../../utils/logger');
+const { logger } = require('@dictadata/lib')
 const { Storage } = require('@dictadata/storage-tracts');
 const { StorageResults, StorageError } = require("@dictadata/storage-junctions/types");