+++ storage/node/node-api/storage.js
@@ -7,10 +7,10 @@ const express = require("express");
 const authorize = require("../authorize");
 const Roles = require("../roles");
 const config = require("../config.js");
-const logger = require('../../utils/logger');
+const { logger } = require('@dictadata/lib')
 const { Storage } = require('@dictadata/storage-tracts');
 const { StorageResults, StorageError } = require('@dictadata/storage-junctions/types');
-const { objCopy } = require('@dictadata/storage-junctions/utils');
+const { objCopy } = require('@dictadata/lib');
 
 /**
  * storage routes