+++ storage/node/node-api/tracts.js
@@ -6,7 +6,7 @@
 const express = require("express");
 const authorize = require("../authorize");
 const Roles = require("../roles");
-const logger = require('../../utils/logger');
+const { logger } = require('@dictadata/lib')
 const { Storage } = require('@dictadata/storage-tracts');
 const { StorageError } = require('@dictadata/storage-junctions/types');