+++ storage/node/node-api/user.js
@@ -8,7 +8,7 @@ const accounts = require("../accounts");
 const Account = require('../account');
 const authorize = require("../authorize");
 const Roles = require("../roles");
-const logger = require("../../utils/logger");
+const { logger } = require('@dictadata/lib')
 const { StorageResults, StorageError } = require("@dictadata/storage-junctions/types");
 
 /**