+++ storage/node/startup.js
@@ -3,7 +3,7 @@
  */
 "use strict";
 
-const { exitOnError } = require("winston");
+// const { exitOnError } = require("winston");
 
 var startup_list = [];